builder: refactor map/array syntax handling routines into generic utilities.
